hardware/requirements.txt
flask>=2.3.0
pyserial>=3.5
python-dotenv>=1.0.0
psutil>=5.9.0
lgpio>=0.2.0.0
numpy>=1.24.0 